Automate updating of Spring Boot starter project
Add script to be run in Travis CI cron build that downloads the latest
Spring Boot starter project with web and actuator starters from
start.spring.io and uses its contents to update this project.  Only
push updates when on master branch.  Make this project look enough
like that stock starter project so this thing can work.  We use
xmllint, so make sure Travis has that installed.

Travis CI puts git in a detached HEAD state which makes committing
changes trickier.  Make sure the HEAD ref and branch ref are the same
and then checkout branch.  If they are not the same, it is not
necessarily a failure, a new push may have come in prior to the build
cloning the repo.

Presently the script will not exit when in a non-cron build for
testing purposes.

Closes #1